refine: ui部分を使いまわせる様に分割 (#35)
